requirements.txt
numpy
scipy
pygame
argparse
pyinstaller
ffmpeg-python
auto-py-to-exe